refactor server_status_monitor.php 1. split long functions 2. add server_status_monitor.lib.php
